Rename files inside the uast folder to remove 'uast_' from the file name
