+++ ml_core/algorithms/__init__.py
@@ -1,10 +1,10 @@
 # flake8: noqa
 from ml_core.algorithms.tf_idf import log_tf_log_idf
 from ml_core.algorithms.token_parser import TokenParser, NoopTokenParser
-from ml_core.algorithms.uast.uast_ids_to_bag import UastIds2Bag, uast2sequence
-from ml_core.algorithms.uast.uast_struct_to_bag import UastRandomWalk2Bag, UastSeq2Bag
-from ml_core.algorithms.uast.uast_inttypes_to_nodes import Uast2QuantizedChildren
-from ml_core.algorithms.uast.uast_inttypes_to_graphlets import Uast2GraphletBag
-from ml_core.algorithms.uast.uast_to_role_id_pairs import Uast2RoleIdPairs
-from ml_core.algorithms.uast.uast_id_distance import Uast2IdLineDistance, Uast2IdTreeDistance
-from ml_core.algorithms.uast.uast_to_id_sequence import Uast2IdSequence
+from ml_core.algorithms.uast.ids_to_bag import UastIds2Bag, uast2sequence
+from ml_core.algorithms.uast.struct_to_bag import UastRandomWalk2Bag, UastSeq2Bag
+from ml_core.algorithms.uast.inttypes_to_nodes import Uast2QuantizedChildren
+from ml_core.algorithms.uast.inttypes_to_graphlets import Uast2GraphletBag
+from ml_core.algorithms.uast.to_role_id_pairs import Uast2RoleIdPairs
+from ml_core.algorithms.uast.id_distance import Uast2IdLineDistance, Uast2IdTreeDistance
+from ml_core.algorithms.uast.to_id_sequence import Uast2IdSequence
