+++ ml_core/extractors/identifiers.py
@@ -1,6 +1,6 @@
 # TODO (Guillemdb): fix imports
 from ml_core.algorithms.token_parser import NoopTokenParser
-from ml_core.algorithms.uast.uast_ids_to_bag import UastIds2Bag
+from ml_core.algorithms.uast.ids_to_bag import UastIds2Bag
 from ml_core.extractors.bags_extractor import BagsExtractor
 from ml_core.extractors.helpers import register_extractor
