+++ ml_core/extractors/uast_seq.py
@@ -1,5 +1,5 @@
 # TODO (Guillemdb): fix imports
-from ml_core.algorithms.uast.uast_struct_to_bag import UastSeq2Bag
+from ml_core.algorithms.uast.struct_to_bag import UastSeq2Bag
 from ml_core.extractors.bags_extractor import BagsExtractor
 from ml_core.extractors.helpers import filter_kwargs, get_names_from_kwargs, register_extractor
